Move prune classes to BlockStore feature
