+++ src/Stratis.Bitcoin.Features.BlockStore/Pruning/IPrunedBlockRepository.cs
@@ -2,7 +2,7 @@
 using NBitcoin;
 using Stratis.Bitcoin.Utilities;
 
-namespace Stratis.Bitcoin.Features.BlockStore
+namespace Stratis.Bitcoin.Features.BlockStore.Pruning
 {
     /// <summary>
     /// Prunes and compacts the block store database by deleting blocks lower than a certain height and recreating the database file on disk.
